\name{as.population_graph}
\alias{as.population_graph}
\title{Converts an object of type matrix or igraph into \code{population_graph}}
\usage{
  as.population_graph(graph)
}
\arguments{
  \item{graph}{An object of type \code{matrix} or
  \code{igraph}}
}
\value{
  An object of type \code{population_graph}
}
\description{
  This is a simple conversion routine for \code{matrix} or
  \code{igraph} objects into \code{population_graph}
  objects
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

