% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_leaflet.R
\name{for_leaflet}
\alias{for_leaflet}
\title{This function convers the popgraph object into an object that can be plot using
 the leaflet library.}
\usage{
for_leaflet(x)
}
\arguments{
\item{x}{A population graph}
}
\value{
A \code{data.frame} object suitable for passing to \code{leaflet}.
}
\description{
This function convers the popgraph object into an object that can be plot using
 the leaflet library.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
