% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_popgraph.R
\name{overlay_popgraph}
\alias{overlay_popgraph}
\title{Convience function to plot edges from \code{popgraph}}
\usage{
overlay_popgraph(graph)
}
\arguments{
\item{graph}{An object of type \code{popgraph}}
}
\description{
This is a quick convienence function for plotting nodes and edges 
  on a normal R plot by iterating through edges and connecting
  nodes.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
