\name{overlay_population_graph}
\alias{overlay_population_graph}
\title{Convience function to plot edges from \code{population_graph}}
\usage{
  overlay_population_graph(graph)
}
\arguments{
  \item{graph}{An object of type \code{population_graph}}
}
\value{
  NULL
}
\description{
  This is a quick convienence function for plotting nodes
  and edges on a normal R plot by iterating through edges
  and connecting nodes.
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

