% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgraph-package.R
\name{popgraph-package}
\alias{popgraph-package}
\title{Package for the creation of population graph objects}
\description{
popgraph is a pacakge that is designed to create and manipulate
 population graph objects (Dyer & Nason 2004, MolEcol).  This 
 specific package was made in conjunction with the book 
 "Applied Landscape Genetics" by R.J. Dyer.
}
\details{
\tabular{ll}{
Package: \tab popgraph\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-06\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

There are some very good examples of the components of this package are used
 in the vignettes for this package.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
\keyword{package}
