% Generated by roxygen2 (4.0.1): do not edit by hand
\name{to_json}
\alias{to_json}
\title{Converts to json representation}
\usage{
to_json(graph, file)
}
\arguments{
\item{graph}{An object of type \code{igraph}}

\item{file}{A file to write the output to (default missing)}
}
\value{
A textual json representation of the graph
}
\description{
This is a nice function that takes a graph structure
 and converts it to a json format for use on the web.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

